+++ packages/axios-extension/test/abap/abap-adt-service.test.ts
@@ -345,6 +345,7 @@ describe('Use existing connection session', () => {
         nock.cleanAll();
         attachUaaAuthInterceptorSpy.mockRestore();
         Uaa.prototype.getAccessToken = jest.fn();
+        Uaa.prototype.getAccessTokenWithClientCredentials = jest.fn();
     });
 
     afterAll(() => {
@@ -369,6 +370,8 @@ describe('Use existing connection session', () => {
         expect(provider.cookies.toString()).toBe('sap-usercontext=sap-client=100; SAP_SESSIONID_Y05_100=abc');
         expect(await provider.isS4Cloud()).toBe(false);
         expect(attachUaaAuthInterceptorSpy).toBeCalledTimes(0);
+        expect(Uaa.prototype.getAccessToken).toBeCalledTimes(0);
+        expect(Uaa.prototype.getAccessTokenWithClientCredentials).toBeCalledTimes(0);
     });
 
     test('abap service provider for cloud - require authentication', async () => {
@@ -385,6 +388,8 @@ describe('Use existing connection session', () => {
         const provider = createForAbapOnCloud(cloneObj as any);
         expect(await provider.isS4Cloud()).toBe(true);
         expect(await provider.user()).toBe('emailTest');
+        expect(Uaa.prototype.getAccessToken).toBeCalledTimes(2);
+        expect(Uaa.prototype.getAccessTokenWithClientCredentials).toBeCalledTimes(0);
     });
 
     test('abap service provider for cloud - with authentication provided', async () => {
@@ -409,6 +414,8 @@ describe('Use existing connection session', () => {
         const provider = createForAbapOnCloud(configForAbapOnCloudWithAuthentication as any);
         expect(await provider.isS4Cloud()).toBe(false);
         expect(await provider.user()).toBe('email');
+        expect(Uaa.prototype.getAccessToken).toBeCalledTimes(0);
+        expect(Uaa.prototype.getAccessTokenWithClientCredentials).toBeCalledTimes(2);
     });
 
     it.each([