+++ packages/control-property-editor/src/panels/index.ts
@@ -1,0 +1,2 @@
+export { PropertiesPanel } from './properties';
+export { LeftPanel } from './LeftPanel';