+++ packages/preview-middleware-client/src/cpe/init.ts
@@ -1,7 +1,72 @@
-import log from 'sap/base/Log';
+import type { ExternalAction } from '@sap-ux-private/control-property-editor-common';
+import {
+    startPostMessageCommunication,
+    iconsLoaded,
+    enableTelemetry
+} from '@sap-ux-private/control-property-editor-common';
 import type RuntimeAuthoring from 'sap/ui/rta/RuntimeAuthoring';
 
-export default function(rta: RuntimeAuthoring) {
-    // custom rta plugin code goes here
-    log.debug(`Editor start with following settings: ${rta.getFlexSettings()}}`);
+import type { ActionHandler, Service } from './types';
+import { initOutline } from './outline/index';
+import { createUi5Facade } from './facade';
+import { SelectionService } from './selection';
+import { ChangeService } from './changes/service';
+import { loadDefaultLibraries } from './documentation';
+import Log from 'sap/base/Log';
+import { logger } from './logger';
+
+
+export default function init(rta: RuntimeAuthoring): Promise<void> {
+    Log.info('Initializing Control Property Editor');
+
+    // enable telemetry if requested
+    const flexSettings = rta.getFlexSettings();
+    if (flexSettings.telemetry === true) {
+        enableTelemetry();
+    }
+
+    const ui5 = createUi5Facade();
+    const actionHandlers: ActionHandler[] = [];
+    /**
+     *
+     * @param handler action handler
+     */
+    function subscribe(handler: ActionHandler): void {
+        actionHandlers.push(handler);
+    }
+
+    const selectionService = new SelectionService(rta, ui5);
+
+    const changesService = new ChangeService({ rta }, ui5, selectionService);
+    const services: Service[] = [selectionService, changesService];
+    try {
+        loadDefaultLibraries();
+        const { sendAction } = startPostMessageCommunication<ExternalAction>(
+            window.parent,
+            async function onAction(action) {
+                for (const handler of actionHandlers) {
+                    try {
+                        await handler(action);
+                    } catch (error) {
+                        Log.error('Handler Failed: ', error);
+                    }
+                }
+            },
+            logger
+        );
+
+        for (const service of services) {
+            service.init(sendAction, subscribe);
+        }
+        // For initOutline to complete the RTA needs to already running (to access RTA provided services).
+        // That can only happen if the plugin initialization has completed.
+        initOutline(rta, ui5, sendAction).catch((error) =>
+            Log.error('Error during initialization of Control Property Editor', error)
+        );
+        const icons = ui5.getIcons();
+        sendAction(iconsLoaded(icons));
+    } catch (error) {
+        Log.error('Error during initialization of Control Property Editor', error);
+    }
+    return Promise.resolve();
 }