+++ packages/preview-middleware-client/test/__mock__/sap/ui/base/DataType.ts
@@ -1,4 +1,10 @@
-// add required functionality for testing here
-export default {
-    getType: jest.fn().mockReturnValue({ getName: () => 'string' })
-};
+export const getNameMock = jest.fn();
+
+export default class DataTypeMock {
+    getName(): string | undefined {
+        return getNameMock();
+    }
+    static getType() {
+        return new DataTypeMock();
+    }
+}