+++ packages/preview-middleware-client/test/__mock__/sap/ui/dt/OverlayUtil.ts
@@ -1,0 +1,5 @@
+import { mockOverlay } from './OverlayRegistry';
+
+export default {
+    getClosestOverlayFor: jest.fn().mockReturnValue(mockOverlay)
+};