+++ packages/preview-middleware/src/base/flp.ts
@@ -10,6 +10,16 @@ import { deleteChange, readChanges, writeChange } from './flex';
 import type { MiddlewareUtils } from '@ui5/server';
 import type { Manifest, UI5FlexLayer } from '@sap-ux/project-access';
 
+const DEVELOPER_MODE_CONFIG = new Map([
+    // Run application in design time mode
+    // Adds bindingString to BindingInfo objects. Required to create and read PropertyBinding changes
+    ['xx-designMode', 'true'],
+    // In design mode, the controller code will not be executed by default, which is not desired in our case, so we suppress the deactivation
+    ['xx-suppressDeactivationOfControllerCode', 'true'],
+    // Make sure that XML preprocessing results are correctly invalidated
+    ['xx-viewCache', 'false']
+]);
+
 /**
  * Enhanced request handler that exposes a list of endpoints for the cds-plugin-ui5.
  */
@@ -70,6 +80,7 @@ export interface TemplateConfig {
             writeConnector: string;
             custom: boolean;
         }[];
+        bootstrapOptions: string;
         resources: Record<string, string>;
     };
     flex?: {
@@ -141,7 +152,8 @@ export class FlpSandbox {
                 resources: {
                     ...resources,
                     [PREVIEW_URL.client.ns]: PREVIEW_URL.client.url
-                }
+                },
+                bootstrapOptions: ''
             },
             locateReuseLibsScript: this.config.libs ?? (await this.hasLocateReuseLibsScript())
         };
@@ -155,6 +167,7 @@ export class FlpSandbox {
         if (this.rta) {
             this.rta.options ??= {};
             this.rta.options.baseId = componentId ?? manifest['sap.app'].id;
+            this.rta.options.generator = 'changeUtils: SAPFioriTools-propertyEditor';
             this.addEditorRoutes(this.rta);
         }
         this.addRoutesForAdditionalApps();
@@ -168,6 +181,7 @@ export class FlpSandbox {
      * @param rta runtime authoring configuration
      */
     private addEditorRoutes(rta: RtaConfig) {
+        const cpe = dirname(require.resolve('@sap-ux/control-property-editor-sources'));
         for (const editor of rta.editors) {
             let previewUrl = editor.path;
             if (editor.developerMode) {
@@ -176,10 +190,16 @@ export class FlpSandbox {
                 this.router.get(editor.path, (async (_req: Request, res: Response) => {
                     const template = readFileSync(join(__dirname, '../../templates/flp/editor.html'), 'utf-8');
                     const html = render(template, {
-                        previewUrl: `${previewUrl}?sap-ui-xx-viewCache=false&fiori-tools-rta-mode=forAdaptation&sap-ui-rta-skip-flex-validation=true&sap-ui-xx-condense-changes=true#${this.config.intent.object}-${this.config.intent.action}`
+                        previewUrl: `${previewUrl}?sap-ui-xx-viewCache=false&fiori-tools-rta-mode=forAdaptation&sap-ui-rta-skip-flex-validation=true&sap-ui-xx-condense-changes=true#${this.config.intent.object}-${this.config.intent.action}`,
+                        telemetry: rta.options?.telemetry ?? false
                     });
                     res.status(200).contentType('html').send(html);
                 }) as RequestHandler);
+                let path = dirname(editor.path);
+                if (!path.endsWith('/')) {
+                    path = `${path}/`;
+                }
+                this.router.use(`${path}editor`, serveStatic(cpe));
             }
             this.router.get(previewUrl, (async (_req: Request, res: Response) => {
                 const config = { ...this.templateConfig };
@@ -189,6 +209,9 @@ export class FlpSandbox {
                     developerMode: editor.developerMode === true,
                     pluginScript: editor.pluginScript
                 };
+                if (editor.developerMode === true) {
+                    config.ui5.bootstrapOptions = serializeUi5Configuration(DEVELOPER_MODE_CONFIG);
+                }
                 const template = readFileSync(join(__dirname, '../../templates/flp/sandbox.html'), 'utf-8');
                 const html = render(template, config);
                 res.status(200).contentType('html').send(html);
@@ -253,12 +276,7 @@ export class FlpSandbox {
      * @returns template configuration for flex.
      */
     private createFlexHandler(): TemplateConfig['ui5']['flex'] {
-        const workspaceConnectorPath = '/preview/WorkspaceConnector';
-        this.router.get(`/resources${workspaceConnectorPath}.js`, (_req: Request, res: Response) => {
-            res.status(200)
-                .contentType('text/javascript')
-                .send(readFileSync(join(__dirname, '../../templates/flp/workspaceConnector.js'), 'utf-8'));
-        });
+        const workspaceConnectorPath = 'open/ux/preview/client/flp/WorkspaceConnector';
         const api = `${PREVIEW_URL.api}/changes`;
         this.router.use(api, json());
         this.router.get(api, (async (_req: Request, res: Response) => {
@@ -329,4 +347,30 @@ export class FlpSandbox {
             url: app.target
         };
     }
+}
+
+/**
+ * Creates an attribute string that can be added to an HTML element.
+ *
+ * @param attributes map with attributes and their values
+ * @param indent indentation thats inserted before each attribute
+ * @param prefix value that should be added at the start of to all attribute names
+ * @returns attribute string
+ */
+function serializeDataAttributes(attributes: Map<string, string>, indent = '', prefix = 'data'): string {
+    return [...attributes.entries()]
+        .map(([name, value]) => {
+            return `${indent}${prefix}-${name}="${value}"`;
+        })
+        .join('\n');
+}
+
+/**
+ * Creates an attribute string that can be added to bootstrap script in a HTML file.
+ *
+ * @param config ui5 configuration options
+ * @returns attribute string
+ */
+function serializeUi5Configuration(config: Map<string, string>): string {
+    return '\n' + serializeDataAttributes(config, '        ', 'data-sap-ui');
 }