+++ packages/system-access/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/system-access
 
+## 0.2.8
+
+### Patch Changes
+
+-   Updated dependencies [0760c9f8]
+    -   @sap-ux/axios-extension@1.4.8
+
 ## 0.2.7
 
 ### Patch Changes